requirements.txt
alembic==1.12.1
annotated-types==0.6.0
attrs==23.1.0
blinker==1.7.0
cachelib==0.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.36.0
cryptography==41.0.7
Deprecated==1.2.14
flasgger==0.9.7.1
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Script==0.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Mako==1.3.0
MarkupSafe==2.1.3
mistune==3.0.2
oauthlib==3.2.2
packaging==23.2
psycopg2==2.9.9
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.14.3
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.0
six==1.16.0
SQLAlchemy==2.0.23
stripe==7.9.0
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
vonage==3.12.0
vonage-jwt==1.1.0
Werkzeug==3.0.1
wrapt==1.16.0
